+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## \[0.3.0\] \[2020-03-22\]
+
+### Added
+
+* More examples.
+
+### Changed
+
+* A lot of changes to package structure. Might be documented here more later.
+
 ## \[0.2.4\] \[2020-03-13\]
 
 ### Added